# 🚀 MAJOR: Fix critical performance bottlenecks in load operations