--- a/PERFORMANCE_OPTIMIZATION_PRD.md
+++ b/PERFORMANCE_OPTIMIZATION_PRD.md
@@ -1,0 +1,181 @@
+# 🚀 DataSON Performance Optimization PRD
+
+## 📋 Executive Summary
+
+Based on comprehensive profiling analysis, DataSON shows critical performance bottlenecks that require immediate investigation. While core serialization is fast (~89ms), security validation (`limits_prepare`) is causing **25+ second delays** - a 280x performance penalty that makes the library unusable for production workloads.
+
+## 🔍 Profiling Findings
+
+### Performance Baseline (Python-only, 1k data)
+| Stage | Median Time | P95 Time | Status |
+|-------|-------------|----------|---------|
+| `serialize_inner_python` | 89ms | 3,526ms | ✅ Acceptable |
+| `eligibility_check` | 291ms | 1,815ms | ⚠️ Could optimize |
+| `limits_prepare` | **25,417ms** | **47,508ms** | 🚨 Critical issue |
+| `save_string` | 59ms | 35,241ms | ⚠️ High variance |
+| `load_basic_json` | 762ms | 6,841ms | ⚠️ Slow for 1k data |
+| `smart_scalars` | 1.8ms | 35ms | ✅ Fast |
+| `postprocess` | 68ms | 1,347ms | ✅ Acceptable |
+
+### Throughput Analysis
+- **Current**: ~2,870 ops/sec
+- **Target with fixes**: 15,000+ ops/sec
+- **Bottleneck**: Security validation blocking all operations
+
+## 🎯 Investigation Priorities
+
+### P0 - Critical (Immediate)
+**1. `limits_prepare` Stage Investigation**
+- **Current**: 25+ second delays
+- **Expected**: <10ms for security checks
+- **Investigate**:
+  - `/datason/_security.py` or similar security modules
+  - Circular imports or infinite loops
+  - Regex performance in validation
+  - File I/O operations in limits checking
+
+**2. Performance Regression Analysis**
+- **Action**: Compare with previous versions
+- **Target**: Identify when this bottleneck was introduced
+- **Files**: Git blame on security/limits related code
+
+### P1 - High Priority
+**3. `load_basic_json` Optimization**
+- **Current**: 762ms for 1k JSON parsing
+- **Expected**: <50ms for 1k data
+- **Investigate**:
+  - JSON parsing implementation in `/datason/api.py`
+  - Smart scalar detection overhead
+  - Type inference algorithms
+
+**4. `eligibility_check` Performance**
+- **Current**: 291ms
+- **Expected**: <10ms
+- **Investigate**:
+  - Data type inspection logic
+  - Complex object traversal
+  - Recursive analysis depth
+
+### P2 - Medium Priority
+**5. Performance Variance Investigation**
+- **Issue**: P95 times 10-100x higher than medians
+- **Investigate**: Memory allocation patterns, GC pressure
+- **Target**: Consistent sub-100ms performance
+
+## 🔍 Code Areas to Investigate
+
+### Primary Investigation Targets
+
+**1. Security/Limits Module**
+```python
+# Likely locations:
+datason/_security.py
+datason/config.py (limit configurations)
+datason/_profiling.py (limits_prepare stage)
+```
+
+**2. Core API Implementation**
+```python
+# Performance-critical paths:
+datason/api.py (dumps_json, loads_json functions)
+datason/__init__.py (save_string, load_basic)
+```
+
+**3. Type System and Validation**
+```python
+# Eligibility and type checking:
+datason/_types.py
+datason/_validation.py
+datason/smart_scalars.py
+```
+
+### Investigation Methodology
+
+**1. Code Path Analysis**
+- Trace execution through profiled stages
+- Identify exact functions called in each stage
+- Look for nested loops, recursive calls, or blocking I/O
+
+**2. Performance Benchmarking**
+- Create isolated tests for each stage
+- Use `cProfile` and `line_profiler` for detailed analysis
+- Test with different data sizes (1k, 10k, 100k)
+
+**3. Regression Testing**
+- Test against previous DataSON versions
+- Identify commit that introduced the bottleneck
+- Compare against standard `json` library performance
+
+## 🎯 Success Metrics
+
+### Performance Targets
+| Stage | Current | Target | Improvement |
+|-------|---------|--------|-------------|
+| `limits_prepare` | 25,417ms | <10ms | **2,541x** |
+| `load_basic_json` | 762ms | <50ms | **15x** |
+| `eligibility_check` | 291ms | <10ms | **29x** |
+| **Overall throughput** | 2,870 ops/sec | 15,000+ ops/sec | **5x** |
+
+### Quality Metrics
+- ✅ P95 latency within 3x of median
+- ✅ Linear scaling with data size
+- ✅ Memory usage under 2x of raw data size
+- ✅ Zero performance regressions in CI
+
+## 🚧 Implementation Plan
+
+### Phase 1: Critical Bug Fixes (Week 1)
+1. **Emergency Investigation**: `limits_prepare` bottleneck
+2. **Root Cause Analysis**: Find and fix 25+ second delays
+3. **Regression Testing**: Ensure no functionality lost
+4. **Performance Validation**: Confirm <10ms security checks
+
+### Phase 2: Core Optimizations (Week 2)
+1. **JSON Parsing**: Optimize `load_basic_json` to <50ms
+2. **Eligibility Checks**: Streamline type validation to <10ms
+3. **Memory Optimization**: Reduce allocation overhead
+4. **Benchmark Suite**: Comprehensive performance testing
+
+### Phase 3: Rust Integration (Week 3-4)
+1. **Baseline Comparison**: Python optimizations vs. Rust
+2. **Selective Acceleration**: Identify highest-impact Rust targets
+3. **Hybrid Architecture**: Python validation + Rust serialization
+4. **Production Validation**: Real-world performance testing
+
+## 🔍 Investigation Questions
+
+### Critical Questions to Answer
+1. **What exactly is `limits_prepare` doing for 25+ seconds?**
+2. **Is this a recent regression or long-standing issue?**
+3. **Why is JSON parsing 15x slower than standard library?**
+4. **Are security checks necessary for every operation?**
+5. **Can validation be cached or optimized?**
+
+### Technical Deep Dive
+1. **Code Path Tracing**: Which functions are called in each stage?
+2. **Resource Usage**: Is this CPU, I/O, or memory bound?
+3. **Concurrency Issues**: Are there locks or blocking operations?
+4. **Data Size Sensitivity**: How does performance scale?
+
+## 🎯 Expected Outcomes
+
+### Immediate (Post Phase 1)
+- ✅ `limits_prepare` under 10ms (from 25+ seconds)
+- ✅ DataSON usable for production workloads
+- ✅ Baseline performance established for Rust comparison
+
+### Medium Term (Post Phase 2)
+- ✅ 5-10x overall performance improvement
+- ✅ Competitive with standard JSON library
+- ✅ Consistent sub-100ms latencies for typical workloads
+
+### Long Term (Post Phase 3)
+- ✅ 10-50x performance with Rust acceleration
+- ✅ Industry-leading serialization performance
+- ✅ Production-ready for high-throughput applications
+
+---
+
+**Priority**: 🚨 Critical - Security validation bottleneck blocking all usage
+**Timeline**: Emergency fix needed within 1 week
+**Owner**: Development team with profiling/performance expertise