@danielendler 2025-08-16
# 🚀 DataSON PR Performance Analysis

## 📊 Performance Metrics (Actual Numbers)

### System Configuration
- **DataSON Version**: 0.12.0
- **Rust Acceleration**: 🐍 Python-only
- **Profiling**: Enabled with nanosecond precision

### 🎯 Basic Performance Test
| Metric | Value |
|--------|-------|
| Dataset Size | 0 chars |
| Serialization Time | **0.00ms** |
| Deserialization Time | **0.00ms** |
| JSON Output Size | 0 chars |
| Profile Events (Save) | 0 |
| Profile Events (Load) | 0 |

### 📈 Performance Scenarios
| Scenario | Size (chars) | Save (ms) | Load (ms) | Total Events | Throughput (ops/s) |
|----------|-------------|-----------|-----------|--------------|-------------------|

## ✅ CI Validation Status
- ✅ All profiling tests passed
- ✅ Round-trip validation successful
- ✅ Performance metrics captured with precision
- ✅ Stage-level timing analysis complete

## 📝 Performance Analysis


---
*This analysis was automatically generated by DataSON's integrated profiling system*
*All metrics are actual measured values from the CI run*


## 📦 Additional Benchmark Data
Full benchmark suite results are available in the workflow artifacts.