@danielendler 2025-08-16
# 🚀 DataSON PR Performance Analysis

## 📊 Performance Metrics (Actual Numbers)

### System Configuration
- **DataSON Version**: 0.12.0
- **Rust Acceleration**: 🐍 Python-only
- **Profiling**: Enabled with nanosecond precision

### 🎯 Basic Performance Test
| Metric | Value |
|--------|-------|
| Dataset Size | 383 chars |
| Serialization Time | **0.15ms** |
| Deserialization Time | **2.97ms** |
| JSON Output Size | 4 chars |
| Profile Events (Save) | 0 |
| Profile Events (Load) | 0 |

### 📈 Performance Scenarios
| Scenario | Size (chars) | Save (ms) | Load (ms) | Total Events | Throughput (ops/s) |
|----------|-------------|-----------|-----------|--------------|-------------------|
| Tiny JSON | 1 | **0.10** | **0.17** | 0 | 3703.7 |
| Small Nested | 2 | **0.06** | **0.67** | 0 | 1369.9 |
| Medium Array | 2 | **0.51** | **50.08** | 0 | 19.8 |
| Large Complex | 2 | **2.52** | **401.74** | 0 | 2.5 |

## ✅ CI Validation Status
- ✅ All profiling tests passed
- ✅ Round-trip validation successful
- ✅ Performance metrics captured with precision
- ✅ Stage-level timing analysis complete

## 📝 Performance Analysis

### Key Findings:
- **Scaling**: 2x data size increase results in 2.7x processing time increase
- **Largest dataset**: 2 chars processed in 0.73ms total
- **Best throughput**: 3703.7 ops/s on Tiny JSON

---
*This analysis was automatically generated by DataSON's integrated profiling system*
*All metrics are actual measured values from the CI run*


## 📦 Additional Benchmark Data
Full benchmark suite results are available in the workflow artifacts.