@danielendler 2025-08-16
# 🚀 DataSON PR Performance Analysis

## 📊 Performance Metrics (Actual Numbers)

### System Configuration
- **DataSON Version**: 0.12.0
- **Rust Acceleration**: 🐍 Python-only
- **Profiling**: Enabled with nanosecond precision

### 🎯 Basic Performance Test
| Metric | Value |
|--------|-------|
| Dataset Size | 383 chars |
| Serialization Time | **0.15ms** |
| Deserialization Time | **2.98ms** |
| JSON Output Size | 4 chars |
| Profile Events (Save) | 0 |
| Profile Events (Load) | 0 |

### 📈 Performance Scenarios
| Scenario | Size (chars) | Save (ms) | Load (ms) | Total Events | Throughput (ops/s) |
|----------|-------------|-----------|-----------|--------------|-------------------|
| Tiny JSON | 1 | **0.10** | **0.16** | 0 | 3846.2 |
| Small Nested | 2 | **0.04** | **0.70** | 0 | 1351.4 |
| Medium Array | 2 | **0.51** | **49.71** | 0 | 19.9 |
| Large Complex | 2 | **2.27** | **401.86** | 0 | 2.5 |

## ✅ CI Validation Status
- ✅ All profiling tests passed
- ✅ Round-trip validation successful
- ✅ Performance metrics captured with precision
- ✅ Stage-level timing analysis complete

## 📝 Performance Analysis

### Key Findings:
- **Scaling**: 2x data size increase results in 2.8x processing time increase
- **Largest dataset**: 2 chars processed in 0.74ms total
- **Best throughput**: 3846.2 ops/s on Tiny JSON

---
*This analysis was automatically generated by DataSON's integrated profiling system*
*All metrics are actual measured values from the CI run*


## 📦 Additional Benchmark Data
Full benchmark suite results are available in the workflow artifacts.